some fixes for the external networking for tcp with --external.  Also some fixes recommended by MSVC2019 on the code analysis.

+++ src/helics/shared_api_library/helicsExport.cpp
@@ -226,7 +226,7 @@ void helicsFederateInfoLoadFromArgs (helics_federate_info fi, int argc, const ch
     try
     {
         std::vector<std::string> args;
-        args.reserve (argc - 1);
+        args.reserve (static_cast<size_t>(argc) - 1);
         for (int ii = argc - 1; ii > 0; --ii)
         {
             args.emplace_back (argv[ii]);
@@ -576,7 +576,7 @@ helics_core helicsCreateCoreFromArgs (const char *type, const char *name, int ar
     {
         core->valid = coreValidationIdentifier;
         std::vector<std::string> args;
-        args.reserve (argc - 1);
+        args.reserve (static_cast<size_t>(argc) - 1);
         for (int ii = argc - 1; ii > 0; ii--)
         {
             args.emplace_back (argv[ii]);
@@ -705,7 +705,7 @@ helics_broker helicsCreateBrokerFromArgs (const char *type, const char *name, in
     try
     {
         std::vector<std::string> args;
-        args.reserve (argc - 1);
+        args.reserve (static_cast<size_t>(argc) - 1);
         for (int ii = argc - 1; ii > 0; ii--)
         {
             args.emplace_back (argv[ii]);
